+++ admin/tabs/activity/tracking/ActivityTrackingModel.js
@@ -231,7 +231,7 @@ export class ActivityTrackingModel extends HoistModel {
 
     async loadGridAndChartAsync() {
         const {cube, gridModel, chartsModel, dimensions} = this,
-            data = cube.executeQuery({
+            data = await cube.executeQueryAsync({
                 dimensions,
                 filter: this.filterChooserModel.value,
                 includeLeaves: true
@@ -240,7 +240,6 @@ export class ActivityTrackingModel extends HoistModel {
         data.forEach(node => this.separateLeafRows(node));
         gridModel.loadData(data);
         await gridModel.preSelectFirstAsync();
-
 
         chartsModel.setDataAndDims({data, dimensions});
     }